\title{iFes-internal}
\name{iFes-internal}
\alias{CV}
\alias{InitLogloss}
\alias{KFold}
\alias{LOOCV}
\alias{LRCUDA}
\alias{LRMultipleGPU}
\alias{LRSingleGPU}
\alias{LRWithFixedVal}
\alias{LRWithFixedValMG}
\alias{RmDupVal}
\alias{cal.f.ratio}
\alias{combineResult}
\alias{generateIds}
\alias{getGPUCount}
\alias{logloss}



\description{
Internal iFes functions
}

\details{
These are not to be called by the user.
}
\keyword{internal}


